Build package in an environment
This is the same issue that's affecting https://github.com/csdms/pymt/pull/27;
the Miniconda3 installer is causing package conflicts. Here, instead of
reverting to the Miniconda2 installer, I'll see if building in an environment
sidesteps the package conflicts.

This only a problem on Travis CI, not on Circle CI.